Moved all heaps into a DataStructures.Heap namespace
Also started documentation on Fheaps overall structure
